Splits the UserList into a presentational List component and a domain specific UserList container
